lista 1 finalizada

+++ exercicios/lista1/questao10.hs
@@ -1,20 +1,20 @@
 import Text.Printf
 
-letPositive :: Int -> Int
-letPositive n = if (n >=0) then n else 0
+letPositive :: Double -> Double
+letPositive n = if n >=0 then n else 0
 
-sumArr :: [Int] -> Int
+sumArr :: [Double] -> Double
 sumArr[] = 0
 sumArr (a:x) = a + sumArr x
 
 
 main :: IO ()
-main = do a <- readLn :: IO Int
-          b <- readLn :: IO Int
-          c <- readLn :: IO Int
-          d <- readLn :: IO Int
-          e <- readLn :: IO Int
-          f <- readLn :: IO Int
+main = do a <- readLn :: IO Double
+          b <- readLn :: IO Double
+          c <- readLn :: IO Double
+          d <- readLn :: IO Double
+          e <- readLn :: IO Double
+          f <- readLn :: IO Double
           let allPositives = map letPositive [a, b, c, d, e, f]
           print(sumArr allPositives)
           print $ sumArr allPositives / 6
